Restructure server application
So that it is structured by modules (currently
we only have modules "core", "auth", and "user")

Also adds index points to clean up the imports a bit
