+++ packages/server/src/app/auth/index.js
@@ -1,0 +1,2 @@
+export * from './auth.router.js';
+export * from './middleware/jwt-auth-header-validator.js';
