+++ packages/server/src/app/core/models/errors/bad-request.exception.js
