+++ packages/server/src/app/core/models/errors/http.exception.js
