+++ packages/server/src/app/core/models/errors/internal-server.exception.js
