+++ packages/server/src/app/core/models/errors/not-found.exception.js
