+++ packages/server/src/app/core/services/index.js
@@ -1,0 +1,2 @@
+export * from './config.service.js';
+export * from './database.service.js';
