+++ packages/server/src/app/index.js
@@ -1,0 +1,2 @@
+export * from './app.js';
+export * from './core/services/index.js';
