+++ packages/server/src/app/user/user.validator.js
@@ -1,22 +1,3 @@
-export const loginSchema = {
-  username: {
-    notEmpty: { bail: true, errorMessage: 'Een gebruikersnaam is verplicht.' },
-    isLength: {
-      options: { min: 3 },
-      errorMessage: 'Een gebruikersnaam moet minimaal 3 tekens bevatten.'
-    },
-    escape: true
-  },
-  password: {
-    notEmpty: { bail: true, errorMessage: 'Een wachtwoord is verplicht.' },
-    isLength: {
-      options: { min: 3 },
-      errorMessage: 'Het wachtwoord is te kort. Geef minimaal 3 tekens op.'
-    },
-    escape: true
-  }
-};
-
 export const newUserSchema = {
   username: {
     notEmpty: { bail: true, errorMessage: 'Een gebruikersnaam is verplicht.' },
